Added nginx-proxy to allow routing traffic to containers
I wanted to ensure that I could access the services by
name. I wanted a reverse proxy that automatically
generates the appropriate configuration for Nginx.

nginx-proxy allows us to do just that. It looks for
containers that have a VIRTUAL_HOST environmental
variable and looks at it's exposed ports and routes
traffic to it's IP on that specified port.

I also updated the documentation so it's clear how
to add apps so that we route to them.

A shortcoming of this approach is that we still
need to add host entries (/etc/hosts) for
each app, which is a little annoying, but we'll
fix that later.